lease_markt/requirements.txt
streamlit>=1.32.0
pandas>=2.0.0
plotly>=5.18.0
pillow>=10.0.0
python-dateutil>=2.8.2
